requirements.txt
Flask==1.1.2
flask-redis==0.4.0
pythonping==1.0.16
pytimeparse==1.1.8
PyYAML==5.4.1
redis==3.5.3
speedtest-cli==2.1.3
wheel==0.36.2
click==7.1.2
